CleanUp code
