+++ README.md
@@ -1,2 +1,4 @@
 # Techopolis Golang homework
 Golang Open Source Homework Repository
+
+* To run the program with test data, run script `run_index.sh`
