+++ index/go.mod
@@ -1,0 +1,3 @@
+module com.github.senyast4745/index
+
+go 1.13
