+++ index/main.go
@@ -1,27 +1,25 @@
 package main
 
 import (
-	"bufio"
 	"encoding/json"
 	"fmt"
-	"log"
 	"os"
 	"path/filepath"
 )
 
-const finalDataFile = "./final.txt"
+const finalDataFile = "./final.json"
 
 func main() {
 	folderLocation := os.Args[1]
 	files, err := filePathWalkDir(folderLocation)
 	check(err, "error %e while reading files from directory")
-	m := make(map[string][]int)
+	m := make(map[string][]string)
 	for fn := range files {
 		if words, err := readFileByWords(files[fn]); err != nil {
 			fmt.Printf("error %e while reading data from file %s", err, files[fn])
 		} else {
 			for i := range words {
-				m[words[i]] = append(m[words[i]], fn)
+				m[words[i]] = append(m[words[i]], files[fn])
 			}
 		}
 	}
@@ -47,43 +45,4 @@ func filePathWalkDir(root string) ([]string, error) {
 		return nil
 	})
 	return files, err
-}
-
-func readFileByWords(fn string) ([]string, error) {
-
-	file, err := os.Open(fn)
-	if err != nil {
-		log.Fatalf("error while ")
-		return nil, err
-	}
-	//noinspection GoUnhandledErrorResult
-	defer file.Close()
-	scanner := bufio.NewScanner(file)
-	scanner.Split(bufio.ScanWords)
-	var data []string
-	for scanner.Scan() {
-		data = append(data, scanner.Text())
-	}
-	if err := scanner.Err(); err != nil {
-		return nil, err
-	}
-	return data, nil
-}
-
-func writeDataToFile(str string) error {
-
-	f, err := os.Create(finalDataFile)
-	if err != nil {
-		return err
-	}
-	//noinspection GoUnhandledErrorResult
-	defer f.Close()
-
-	if _, err := f.WriteString(str); err != nil {
-		return err
-	}
-	if err := f.Sync(); err != nil {
-		return err
-	}
-	return nil
 }
