+++ run_index.sh
@@ -1,0 +1,4 @@
+#! /bin/bash +x
+cd ./index || exit
+go build
+./index ./data
